# specs/001-web-dashboard/contracts/api.yaml
openapi: 3.0.3
info:
  title: Web Dashboard API
  version: 0.1.0
servers:
  - url: /api
paths:
...
  /stocks/{ts_code}/daily:
    get:
      summary: 获取股票日线行情
      parameters:
        - in: path
          name: ts_code
          required: true
          schema:
            type: string
        - in: query
          name: start_date
          schema:
            type: string
          description: 起始日期（YYYYMMDD or YYYY-MM-DD）
        - in: query
          name: end_date
          schema:
            type: string
          description: 结束日期（YYYYMMDD or YYYY-MM-DD）
      responses:
        "200":
          description: 日线行情数据
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyKlineResponse"
...
components:
  schemas:
...
    DailyKline:
      type: object
      properties:
        trade_date:
          type: string
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
        amount:
          type: number
        adj_factor:
          type: number
      required: [trade_date, open, high, low, close, volume]
...
    DailyKlineResponse:
      type: object
      properties:
        ts_code:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/DailyKline"
      required: [ts_code, items]
...